refactor(cloudflare): move script into its own file
