+++ infrastructure/cloudflare-workers/curl-landing/index.ts
@@ -1,4 +1,6 @@
 import * as cloudflare from "@pulumi/cloudflare";
+import * as fs from 'fs';
+import * as path from 'path';
 
 const account = new cloudflare.Account('fabien.sh', {
     name: 'fabien.sh'
@@ -10,40 +12,12 @@ const zone = new cloudflare.Zone("fabien.sh", {
     zone: 'fabien.sh',
 }, {protect: true,});
 
+const scriptPath = path.join(__dirname, 'script', 'curl_landing.js');
+
 const script = new cloudflare.WorkerScript('curl-landing', {
-    accountId: account.id,
-    content: `/**
- * Returns whether a request has a User-Agent
- * hinting the requests comes from the curl CLI tool or library
- *
- * @param {Request} request
- * @return {boolean}
- */
-const isCurlUserAgent = (request) => {
-  return request.headers.get('user-agent')?.startsWith('curl/');
-}
-
-export default {
-  async fetch(request, env, ctx) {
-    if (isCurlUserAgent(request)) {
-      return Response.json(
-        {
-          roles: ['cloud-engineer', 'platform-engineer', 'sre'],
-          status: 'freelance',
-          available: true,
-          topics: ['dev', 'infrastructure', 'platform']
-        },
-        {
-          status: 203,
-          statusText: 'Non-Authoritative Information'
-        });
-    }
-
-    return fetch(request);
-  },
-};
-`,
     name: 'curl-landing',
+    accountId: account.id,
+    content: fs.readFileSync(scriptPath, 'utf8'),
     module: true,
 }, {protect: true});
